add "end" at end of function
There was no "end" at the end of the function. I don't understand why it did work before. (after adding the last part, "amplitudes.txt", that part did not work when called from the master, that might have been because of the non-existing end, but the other parts did run from the master before (weird!)